Made guest users reusable; conversion between guest and registered users completed
Guest users are now reusable. The application checks if a guest user already exists before creating a new one.

For converting a guest to a registered user, a new endpoint has been created. Authentication requires a valid session, and the current user must have the role "GUEST".

During conversion, the user’s personal data overwrites the existing guest data and the role is updated to "USER"